Allow core to handle status updates on recurring when core is appropriately patched
CRM-17655 passes the responsibility to core for updating the contribution_status & next_sched_contribution_date. Where sites are patched with it IATS should allow core to manage it.

Note that currently in 4.6 when renewing memberships using a recurring credit card payment the contribution_status_id is being set to 'In Progress' & the next_sched date is being set even when an invalid credit card is being used. (also note there are other patches in 4.6.11 relevant to this flow - so use latest 4.6 from github to do any testing).